requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
python-dateutil==2.8.2